unuse f32

+++ src/compositor/mod.rs
@@ -1,5 +1,5 @@
 pub mod normal;
 
 pub trait Compositor<T> {
-    fn composite(&self, a: &T, b: &T, alpha: f32) -> T;
+    fn composite(&self, a: &T, b: &T, alpha: f64) -> T;
 }
